extract BaselineFilter
